third_party/blink/web_tests/fast/hidpi/image-srcset-src-selection-1x-both.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
    }
    addEventListener("load", function() {
        shouldBeTrue('document.getElementById("foo").clientWidth==400');
    }, false);
</script>
<div>This test passes if the div below is a blue 100px square when the deviceScaleFactor is 1.
    It simply ensures that when both src and srcset are specified for the same DPR, srcset wins.</div>
<img id="foo" src="resources/blue-100-px-square.png" srcset="resources/green-400-px-square.png 1x">
